+++ prepare_standalone.py
@@ -9,7 +9,8 @@ branch_name = Repository('.').head.shorthand
 is_main = branch_name == 'main'
 this_file = os.path.basename(__file__)
 script_name = " ".join(this_file.split(".")[0].split("_"))
-parser = ArgumentParser(prog=script_name, usage=f"python3 {this_file} <path to app>")
+parser = ArgumentParser(prog=script_name, usage=f"python3 {
+                        this_file} <path to app>")
 parser.add_argument("-i", action="store", help=".app path")
 args = parser.parse_args()
 
@@ -29,7 +30,8 @@ def replace_file(source_file, destination_file):
         raise FileNotFoundError(f"Source file '{source_file}' does not exist.")
 
     if not os.path.exists(destination_file):
-        raise FileNotFoundError(f"Destination file '{destination_file}' does not exist.")
+        raise FileNotFoundError(f"Destination file '{
+                                destination_file}' does not exist.")
 
     try:
         shutil.copyfile(source_file, destination_file)
@@ -57,6 +59,7 @@ def edit_max_interface(path: str):
         return test
 
     data["interface"]["menus"] = list(filter(fun, data["interface"]["menus"]))
+
     def keep_save(x):
         keep = x in ["customsave"]
         if not keep:
@@ -74,6 +77,32 @@ def edit_max_interface(path: str):
         json.dump(data, f, indent=4)
 
 
+def get_negative_color(color):
+    # Assuming RGBA, invert the first three components (R, G, B)
+    return [1.0 - c if i < 3 else c for i, c in enumerate(color)]
+
+# Read and modify the JSON file
+
+
+def invert_maxtheme_colors(theme_path):
+
+    with open(theme_path, 'r') as file:
+        data = json.load(file)
+
+    # Modify the colors in the "colors" array
+    if 'colors' in data:
+        for color_entry in data['colors']:
+            if 'oncolor' in color_entry:
+                color_entry['oncolor'] = get_negative_color(
+                    color_entry['oncolor'])
+
+    # Write the modified maxtheme to a new file
+    with open(theme_path, 'w') as file:
+        json.dump(data, file, indent=2)
+
+    print(f"default Max theme colors inverted.")
+
+
 edited = False
 for root, dirs, files in os.walk(app_path):
     folder = os.path.basename('root')
@@ -84,6 +113,8 @@ for root, dirs, files in os.walk(app_path):
             edited = True
             print(f'replacing logo: {file_name}')
             replace_file(logo_path, file_path)
+        if file_name == "default" and file_ext == ".maxtheme":
+            invert_maxtheme_colors(file_path)
         if file_name == "maxinterface":
             edited = True
             edit_max_interface(file_path)